+++ auth/src/index.go
@@ -7,7 +7,7 @@ import (
 
 	"github.com/BeatAllTech/ChitSlip/auth/src/db"
 	"github.com/BeatAllTech/ChitSlip/auth/src/errors"
-	"github.com/BeatAllTech/ChitSlip/auth/src/middleware"
+	"github.com/BeatAllTech/ChitSlip/auth/src/middlewares/currentuser"
 	"github.com/BeatAllTech/ChitSlip/auth/src/routes"
 )
 
@@ -15,7 +15,7 @@ func main() {
 	if os.Getenv("JWT_KEY") == "" {
 		panic("JWT_KEY must be definied")
 	}
-	http.HandleFunc("/api/users/currentuser", middleware.IsLoggedIn(http.HandlerFunc(routes.HandleCurrentUser)))
+	http.Handle("/api/users/currentuser", currentuser.CurrentUser(http.HandlerFunc(routes.HandleCurrentUser)))
 	http.HandleFunc("/api/users/signup", routes.HandleSignUp)
 	http.HandleFunc("/api/users/signout", routes.HandleSignOut)
 	http.HandleFunc("/api/users/signin", routes.HandleSignIn)
