+++ auth/src/routes/currentuser.go
@@ -4,16 +4,19 @@ import (
 	"encoding/json"
 	"fmt"
 	"net/http"
+
+	"github.com/BeatAllTech/ChitSlip/auth/src/middlewares/currentuser"
 )
 
 //HandleCurrentUser handles what happens when the current user visits page
 func HandleCurrentUser(res http.ResponseWriter, req *http.Request) {
 
-	key := "currentUser"
-	if user := req.Context().Value(key); user != nil {
+	if user := req.Context().Value(currentuser.Key); user != nil {
 		data, _ := json.Marshal(user)
 		fmt.Fprintf(res, string(data))
+	} else {
+		fmt.Println("Key not found: ", currentuser.Key)
+
 	}
-	fmt.Println("key not found: ", key)
 
 }
